Refactor and fix measurement layer
